backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# LangGraph and AI
langgraph==0.2.45
langchain-core==0.2.43
langchain-openai==0.1.25
openai==1.40.0
tiktoken==0.8.0

# RSS and Web Scraping
feedparser==6.0.11
beautifulsoup4==4.12.3
trafilatura==1.8.0
aiohttp==3.9.3
urllib3==2.1.0

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
pydantic==2.10.3
pydantic-settings==2.1.0

# Rate limiting
aiolimiter==1.1.0
